Use observation_date for firefox_nondesktop_mau28_*
The 2019 goal definition of mau has defined an offset such that MAU for
2019-01-29 should consider the previous 28 days not including the current day,
so 2019-01-01 through 2019-01-28.

This PR changes our nondesktop MAU tables to call this field `observation_date`
and puts `submission_date` into a `last_submission_date_in_window` field to
hopefully make this self-explanatory without needing to look up documentation.

If we make this change, I'll follow up with similar changes for the desktop
tables.