Add normalized_channel to nondesktop dimensions
For a time, we thought we were going to use release only for KPI
calculations, but this is likely going to revert to the historical
stance of including prerelease in MAU calculations.

This PR removes the filter we had on channel, and retains normalized_channel
as a dimension so we can filter by channel if needed at the analysis level.