Add WAU and id_bucket to desktop mau datasets
Mirrors what we've done for nondesktop.

Actually rolling this out will involve getting the change merged
to `spark-parquet-to-bigquery`, then doing a manual recreation of the table
and downstream. Because there's no windowing or joining involved, though,
those recreations can each be done efficiently in a single query that takes
perhaps a few minutes to run.

This change will allow us to move the Growth Dashboard off of Spark-based
analysis of Parquet datasets and onto BQ (via Redash).