Alter firefox_accounts_* column names to match conventions
As discussed in https://github.com/mozilla/firefox-data-docs/pull/258
the current column name for tier 1 mau in firefox accounts tables
uses a suffix rather than a prefix.

This renames the column in the by_dimensions view rather than in
the raw table so that we avoid having to backfill and update the
Airflow DAG.

In the exact_mau_v1 view, we continue to use the simple "tier1_mau" column
so that we have consistently named columns across the final views
for desktop, nondesktop, and fxa. And these columns are exactly
match exactly with what is presented in the growth dashboard
(we show tier 1 plots for all 3, but only FxA is using the
"seen in tier 1 country" usage criterion).