Join with public dataset for country code lookup
This PR replaces a CASE statement we were using to partially convert
country names to country codes. Rather than maintain our own static
dataset of country data, we join with one of the public datasets
that Google maintains.

This could be a useful pattern for some other common lookups in the future.