Add a test harness for UDFs
This is the first stage for addressing #141

With this change, we parse dependencies between UDFs so that each
UDF sql file can contain just a single UDF definition.
Furthermore, we allow additional SQL statements in the file which
will be treated as tests; they should call ERROR to cause
a test to fail.