Make smoot_usage_all_mtr compatible with scheduled queries
This was erroring because it issues a DDL statement, but our machinery for
running queries sets a destination table. We convert to a normal query so that
the machinery can proceed as normal, populating the destination table.

With this change, it follows the same pattern as the kpi_dashboard table.