Remove default views
We have integrated generation of default views into our schema deployment
pipeline and we have decided that we want to rely on that so that new tables
are automatically presented to users as views without manual intervention in
this repo.

This PR removes all of the views whose text exactly matches what the pipeline
produces as the default. Views like sync ping and core where we've made changes
are not removed here.

Removing the views here should be a no-op as far what the schema propagation
pipeline actually publishes.