Fix bug in udf_js_jackknife_sum_ci for missing values
If we run this function on a grouping where most buckets have no elements,
the function may receive a single-element array. Currently, this leads to
NaN results which throws an exception when passing results back to BigQuery.

Instead, we now infer 0 counts for all missing buckets, which should lead to
more correct results and avoid these errors.