Bug fixes for fenix clients_last_seen query
When doing a `FULL JOIN`, rows that do not have a match will show nulls for
the values of all columns of the missing table, but the table itself will not
pass an `IS NULL` check.

This PR modifies the logic to check `table.client_id IS NOT NULL` instead, and
also makes sure we filter out rows that have not seen a baseline ping in the
relevant window.

With these changes, we should be guaranteed to have at least one populated
day in `baseline.days_seen_bits` and non-null `normalized_channel`.
I am backfilling and validating now.