Add incline exec dash submission_date based query (#843)
* Add incline exec dash submission_date based query

* Move to derived dataset

* Create migrated_clients_v1 as materialized

Previously we had this as a view. Now it will
update every day, taking the latest value seen from migration
pings as its value.

* Update view for materialized query

* Use clients_last_seen view and migrated_clients

Previously, we were using the materialized clients_last_seen
and the unmaterialized migrated_clients. We are now materializing
the latter, and using the view of the former.

- Added documentation
- Removed unecessary arrays

* Run SQL formatter

* Change keyword identifier with prefixed underscore

* Ignore missing table errors in dryrun

* Run migrated_clients init only

* Add docline for negative churn