Fix improper client counting from multiple joins
In this query we are using CROSS JOIN UNNEST to create
extra groups. This is effective only if you are grouping
on those groups, or else those clients end up double counted.

However, we were doing another CROSS JOIN UNNEST to create
more country groups. What this resulted in was extra counting
of the client for every country group they were in. This actually
made non-broken out countries unaffected, but doubled the counts
of broken-out countries (since they have two groups), and ~30%
increase in Overall counts (from the broken out countries).

The fix here is to remove the extra CROSS JOIN UNNEST for country
bucketing, and instead append that array to the country_group
break-out.