Remove queries that write to derived-datasets.telemetry
The kpi_dashboard query is out of date; the 2020 dashboard is implemented in
Databricks and performs a modified version of the query logic.
We remove this table completely and we will send out a fx-data-dev email
to that effect.

Otherwise, the desktop exact mau table was the last scheduled query writing
to the telemetry dataset in the derived-datasets project.