Set country to NULL for unknown search engines
Previously the behavior was to error on unknown search engines
in the revenue join. However this causes failures when we
add new normalized search engines but don't have revenue data
available for them.

Instead we will set the country to NULL, and let this data
fall out during the join with revenue data, which won't
have that country.