Use left joins for amo_stats_dau
Fixes https://bugzilla.mozilla.org/show_bug.cgi?id=1643683

When we introduced Fenix data, we no longer are guaranteed to have values for
each of the per_* aggregations. In particular per_fenix_build will be empty
for any addon that's not compatible with Fenix. We need to use LEFT JOINS to
ensure we don't throw lose addons with no matching rows in per_fenix_build.