Bug 1661250 - prepare Shredder script for adding Pioneer support (#1314)
This adds a new `environment` parameter to Shredder delete script, accepting either `telemetry` or `pioneer` values.
The goal of this parameter is to have a way for choosing different source and target tables. By default it is set to `telemetry` (which keeps current behavior unchanged).

Next step is to define logic for discovering Pioneer study tables in `config.py` and invoke it in delete script (`NotImplementedError`).