Improve ergonomics of creating/scheduling queries with long names
Before, you would get a confusing error about the autogenerated task
name being too long with `./bqetl query schedule`. In theory, you
should have been able to get around this with the `--task` parameter,
but in practice it would autogenerate a name when getting the list
of DAGs, causing a failure that could only be worked around by
manually editing the metadata.

This PR moves the --task-name parameter to query creation and fails
in a (hopefully obvious) way if the autogenerated one would be too long
(if it isn't specified).