Remove jsonPayload.fields.id from fxa_auth query (#1683)
This query failed on the 2021-01-18 run with:

> jsonPayload.fields.id has changed type from STRING to FLOAT

Attempts to run this query in the BQ console show errors because the wildcard
query picks up only the most recent schema, which is incompatible with
historical days. We replace the field with null for now and can investigate
with FxA folks what might have caused this schema change.