Add missing non-versioned telemetry views (#1720)
These were found as part of investigating
https://bugzilla.mozilla.org/show_bug.cgi?id=1671933

They are defined in derived-datasets even though they weren't in bigquery-etl
and so there may be user queries that depend on them. These may have been
erroneously deleted in a previous refactor of this repo.

Restoring them here so that they'll be published to the new `mozdata` project.