Add mobile_usage_2021 for KPI reporting (CDOU)
Mobile counterpart to https://github.com/mozilla/bigquery-etl/pull/1823

Note that we don't yet have `first_seen_date` ETL for mobile, so can't include
`new_profiles` here (which we have for desktop).

This is building on the existing nondesktop exact MAU table, which filters on
`contributes_to_2020_kpi` which so far is identical to the 2021 list, so should
be fine for now.

We may need to rebuild and clean up this ETL pipeline as we improve ETL for
Glean apps in H1.