Bug 1703362 Purge security.unexpectedload events
See https://bugzilla.mozilla.org/show_bug.cgi?id=1703362

The fix that reduces the quantity of events is scheduled to be included
in Firefox 88, released in late April. From 88 on, the rate of events
is expected to be reasonable, and the filter here will allow them again.

I will run a backfill for this table, starting from 2020-12-15.