Bug 1707395 Don't publish views when running glean_usage
Originally, the only way views got published for `baseline_clients_*` tables
was as part of the query runs, but we have since implemented the `output-only`
option and integrated that into `script/generate-sql`, so these views will
get created and updated with normal view publishing runs.

Thus, we no longer need to publish the views along with the queries, which
removes a circular dependency problem we have right now between clients_daily
and clients_first_seen.