Bug 1708264 - Add initial support for dataset_metadata.yaml
This adds the machinery to create `dataset_metadata.yaml` as part of
`bqetl query create` and to validate that these files are valid, but these
are not yet actually used for anything.

The next step will be to fill out `dataset_metadata.yaml` for all
existing derived and user-facing datasets.