Disambiguate ad_clicks in addons_daily query (#2156)
* Disambiguate ad_clicks in addons_daily query

The addons_daily_v1 query failed last night with an error about `ad_clicks`
being an ambiguous field name. See
[Airflow logs](https://workflow.telemetry.mozilla.org/log?execution_date=2021-06-23T04%3A00%3A00%2B00%3A00&task_id=telemetry_derived__addons_daily__v1&dag_id=bqetl_addons).

This is due to https://github.com/mozilla/bigquery-etl/pull/2151 which added
a field called ad_clicks to clients_daily and downstream tables. That new field
is present in a join.

This PR disambiguates the reference by specifying which table to pull it from.

* Allow dry run for search queries