Remove feature_usage_v1
This table was introduced last year as an experiment in
https://github.com/mozilla/bigquery-etl/pull/1193

We never promoted this to a full scheduled query, so the data in the table is
stale. The general idea lives on, however, in the `feature_usage_v2` table.

This PR removes the v1 query, and I will delete the underlying table after
this is merged.