Streamline code to generate events_daily queries and re-run (#2307)
We're still checking the queries into the tree for now, but this moves the code to do so
into bqetl (more discoverable and easier to test) and generally streamlines things so it
will be easier to generate these queries at build-time. Also, check-in updated views from
a run of this code.