Add Windows 10/11 21H2 update to aggregates (#2500)
Windows 11 still has os_version = 10 like Windows 10, so the only way to differentiate is by the build number. Might as well keep them on the same dashboard. This also cleans up older versions which currently combine to ~5% of overall Win10 user share into one grouping to cut down on some of the clutter as the number of releases increases over time.