Bug 1761790 Point to v1 schema in impression_stats view
See https://bugzilla.mozilla.org/show_bug.cgi?id=1761790

We'll be adding a new v2 schema that's not yet used, so explicitly adding
the view definition here so we don't automatically deploy a view that only
looks at v2.

As we build this out, we'll need to make a coordinated change that unions
the two tables and change downstream tables to accept a string-type tile `id`
field.