GLAM ETL: filter out unofficial desktop builds in FOG pipeline (#2957)
GLAM ETL: filter out unofficial desktop builds in FOG pipeline

In https://github.com/mozilla/glam/issues/1941 it was discovered that some desktop clients send version "1024" in FOG telemetry. This is breaking GLAM aggregations because they focus on latest released versions. In order to address that, from now on we'll:
1. filter pings by build_id against official Mozilla builds published in Buildhub. This is equivalent to the filtering we already do in the legacy telemetry GLAM ETL: https://github.com/mozilla/bigquery-etl/blob/9bca48821a8a0d40b1700cc14ecd8068d132ed06/sql/moz-fx-data-shared-prod/telemetry_derived/clients_daily_keyed_scalar_aggregates_v1/query.sql#L3-L8
2. Explicitly filter out pings with version "1024.0.0". At the moment we do not know why this version number is sent by clients (see https://bugzilla.mozilla.org/show_bug.cgi?id=1768187 for details)