Bug 1783427 - GLAM ETL: filter out negative values in aggregate queries
This should fix `extract_probe_counts_v1` query failing due to integer overflows on casting.
We got some big negative numbers in few histogram probes around 2022-08-04 which seem to be some random bit flips. I'm putting a lower bound for a value in a filter at 0 as we do not expect them to be negative in any case.