Update Shredder Pioneer config to account for shared views
Attention stream datasets are shared to other projects.
When analysts create derived datasets from these shared views,
it failed Shredder, which was looking for a PIONEER_ID to delete.

This change will give the correct RALLY_ID to delete, but will not
actually delete the corresponding rows. As a follow-up to this,
we will delete with the appropriate shared-view deletion request.