SVCSE-1595 Convert binary fields to hex strings in FxA database queries
We already convert binary `uid` fields to hex strings in these queries, this does the same with the rest to make querying easier.

Because schemas are changed, after merging this I will manually trigger the queries in Airflow so the next schema deployment doesn't fail. Tables will be overwritten because they are not day-partitioned and queries are ran with `"writeDisposition": "WRITE_TRUNCATE"`.