Pick last expires and types from the history
The history is ordered from first introduction of the metric to the
latest change.
So the last element is the _current_ valid value for expires (and types).

For example this can be seen for `top_sites.contile_reporting_url` in
Firefox Android which initially was set to expire in v112, but has been
change to never expire later. The current database still lists 112 for
it though.