fix: SUM after join error in fxci_derived.task_run_costs_v1
I noticed a SQL error where I was aggregating after joining, causing us
to count worker costs / uptime more than once.

I'm a bit unclear if this was actually causing the data to be wrong or
not because we really only care about the cost / time (and this bug was
overcounting both by the same factor).

But regardless, it's best to fix the problem by doing the aggregation
prior to joining.

Bug: 1926017