MC-1608 Add job for New Tab engagement by corpus_item_id
This job is similar to bqetl_merino_newtab_extract_to_gcs, but it aggregates
on corpus_item_id instead of scheduled_corpus_item_id. The latter (as the name implies)
identifies a scheduled item. We want to start aggregating on the item itself
because we will stop scheduling them for specific dates as part of an experiment
in Q1.