MC-1608 Add job for New Tab engagement by corpus_item_id (#6743)
* MC-1608 Add job for New Tab engagement by corpus_item_id

This job is similar to bqetl_merino_newtab_extract_to_gcs, but it aggregates
on corpus_item_id instead of scheduled_corpus_item_id. The latter (as the name implies)
identifies a scheduled item. We want to start aggregating on the item itself
because we will stop scheduling them for specific dates as part of an experiment
in Q1.

* Revert "MC-1608 Add job for New Tab engagement by corpus_item_id"

This reverts commit f26af987bef0b19f74d47dd28b66a42c89ad8cf1.

* add corpus_item_id to newtab_merino_extract_v1

* simplify query

* Revert adding required-column argument

* Fix format errors

---------

Co-authored-by: Chelsey Beck <64881557+chelseybeck@users.noreply.github.com>