Check snake_casing in object keys for recorded_nimbus_context (#8333)
Glean accidentally defaulted to language-specific casing in the JSON payload from Kotlin & Swift (bug 1931891).
That was fixed in v66 and Glean now sends the `snake_case` version.
This changes the `recorded_nimbus_context` data; for now we can see
which key exists in the JSON object.
Note that the data will never contain both the snake_case and the camelCase version!